+++ workbook-5/exercises/MajesticYardCare/src/main/java/com/pluralsight/models/Square.java
@@ -2,7 +2,7 @@ package com.pluralsight.models;
 
 public class Square extends Rectangle
 {
-    public Square(int sideLength)
+    public Square(double sideLength)
     {
         super("Square", sideLength, sideLength);
     }
